he-he-he-he. That's what they said
A: We just gonna deprecate mediagroup
A: It is no use
A: In fact, It could be done with Javascript

Helloooooooooo~~~~~~~~~~? What are you talking about?

Did you mean wanting to replacing
HTML with JS like that? heeee~~~~?

I wonder how hard is it to just being honest
that you want to replace HTML with JS instead?

Dear god, Look at these big companies sin.
They are quietly wanting to cryptomining in our computer.

Signed-off-by: Yonle <yonle@lecturify.net>